Fix typescript plugin in vscode
This is done by tweaking the logic slightly (typescript 5 returns a
slightly different AST tree) and also use the typescript util
functions that are passed into the plugin. This is important because
the vscode extension currently bundles a typescript version which
might be different from the one that the project is using. And given
that typescript relies on enums with numbers, we can run into issues
where the value 357 for SyntaxKind might mean one thing in one
version and something completely different in another typescript
version.